+++ src/databricks/labs/ucx/contexts/cli_command.py
@@ -118,7 +118,6 @@ class WorkspaceContext(CliContext):
             self.external_locations,
             self.azure_resource_permissions,
             self.azure_resources,
-            self.azure_acl,
             self.principal_acl,
         )
 
@@ -156,6 +155,7 @@ class WorkspaceContext(CliContext):
             self.aws_resources,
             self.external_locations,
             self.inventory_database,
+            self.principal_acl,
             self.named_parameters.get("aws_account_id"),
             self.named_parameters.get("kms_key"),
         )
